Add support for rendering fonts and dynamic text.

+++ LottieUWP/Font.cs
@@ -4,7 +4,7 @@ namespace LottieUWP
 {
     internal class Font
     {
-        private Font(string family, string name, string style, double ascent)
+        private Font(string family, string name, string style, float ascent)
         {
             Family = family;
             Name = name;
@@ -18,7 +18,7 @@ namespace LottieUWP
 
         internal readonly string Style;
 
-        internal readonly double Ascent;
+        internal readonly float Ascent;
 
         internal static class Factory
         {
@@ -27,7 +27,7 @@ namespace LottieUWP
                 string family = json.GetNamedString("fFamily", "");
                 string name = json.GetNamedString("fName", "");
                 string style = json.GetNamedString("fStyle", "");
-                double ascent = json.GetNamedNumber("ascent", 0);
+                float ascent = (float)json.GetNamedNumber("ascent", 0);
                 return new Font(family, name, style, ascent);
             }
         }
